refactor entity-impl data api services
