+++ src/app/api/data-mgt/http/entity-impl/people.http-data-api.service.ts
