+++ src/app/api/data-mgt/local-memory/entity-impl/clients.local-memory-data-api.service.ts
