+++ src/app/api/data-mgt/local-memory/entity-impl/people.local-memory-data-api.service.ts
